+++ src/tests/fixtures/inheritance/classInheritsFieldThroughNonGraphQLIntermediateParent.ts
@@ -1,0 +1,13 @@
+/** @gqlType */
+class Parent {
+  /** @gqlField */
+  parentField: string;
+}
+
+class Intermediate extends Parent {}
+
+/** @gqlType */
+export class Child extends Intermediate {
+  /** @gqlField */
+  childField: string;
+}