+++ src/tests/fixtures/interfaces/ClassInterfaceImplementsInterface.ts
@@ -1,0 +1,18 @@
+/** @gqlInterface Node */
+interface GqlNode {
+  /** @gqlField */
+  id: string;
+}
+
+/** @gqlInterface */
+interface Person extends GqlNode {
+  id: string;
+  /** @gqlField */
+  name: string;
+}
+
+/** @gqlInterface */
+class Actor implements GqlNode, Person {
+  id: string;
+  name: string;
+}